[Decomp][Player] Cohesion split: extract channel methods to PlayerChannel.cpp
