+++ src/game/Object/Player.cpp
@@ -5020,124 +5020,10 @@ void Player::RepopAtGraveyard()
     }
 }
 
-/**
- * @brief Registers a joined chat channel on the player.
- *
- * @param c The channel that was joined.
- */
-void Player::JoinedChannel(Channel* c)
-{
-    m_channels.push_back(c);
-}
 
-/**
- * @brief Removes a chat channel from the player's joined channel list.
- *
- * @param c The channel that was left.
- */
-void Player::LeftChannel(Channel* c)
-{
-    m_channels.remove(c);
-}
 
-/**
- * @brief Leaves and cleans up all channels currently joined by the player.
- */
-void Player::CleanupChannels()
-{
-    while (!m_channels.empty())
-    {
-        Channel* ch = *m_channels.begin();
-        m_channels.erase(m_channels.begin());               // remove from player's channel list
-        ch->Leave(this, false);                             // not send to client, not remove from player's channel list
-        if (ChannelMgr* cMgr = channelMgr(GetTeam()))
-        {
-            cMgr->LeftChannel(ch->GetName()); // deleted channel if empty
-        }
-    }
-    DEBUG_LOG("Player: channels cleaned up!");
-}
 
-/**
- * @brief Updates built-in local channels after a zone change.
- *
- * @param newZone The new area identifier used for localized channel names.
- */
-void Player::UpdateLocalChannels(uint32 newZone)
-{
-    if (m_channels.empty())
-    {
-        return;
-    }
 
-    AreaTableEntry const* current_zone = GetAreaEntryByAreaID(newZone);
-    if (!current_zone)
-    {
-        return;
-    }
-
-    ChannelMgr* cMgr = channelMgr(GetTeam());
-    if (!cMgr)
-    {
-        return;
-    }
-
-    std::string current_zone_name = current_zone->area_name[GetSession()->GetSessionDbcLocale()];
-
-    for (JoinedChannelsList::iterator i = m_channels.begin(), next; i != m_channels.end(); i = next)
-    {
-        next = i; ++next;
-
-        // skip non built-in channels
-        if (!(*i)->IsConstant())
-        {
-            continue;
-        }
-
-        ChatChannelsEntry const* ch = GetChannelEntryFor((*i)->GetChannelId());
-        if (!ch)
-        {
-            continue;
-        }
-
-        if ((ch->flags & 4) == 4)                           // global channel without zone name in pattern
-        {
-            continue;
-        }
-
-        //  new channel
-        char new_channel_name_buf[100];
-        snprintf(new_channel_name_buf, 100, ch->pattern[m_session->GetSessionDbcLocale()], current_zone_name.c_str());
-        Channel* new_channel = cMgr->GetJoinChannel(new_channel_name_buf, ch->ChannelID);
-
-        if ((*i) != new_channel)
-        {
-            new_channel->Join(this, "");                    // will output Changed Channel: N. Name
-
-            // leave old channel
-            (*i)->Leave(this, false);                       // not send leave channel, it already replaced at client
-            std::string name = (*i)->GetName();             // store name, (*i)erase in LeftChannel
-            LeftChannel(*i);                                // remove from player's channel list
-            cMgr->LeftChannel(name);                        // delete if empty
-        }
-    }
-    DEBUG_LOG("Player: channels cleaned up!");
-}
-
-/**
- * @brief Leaves the currently joined looking-for-group channel, if any.
- */
-void Player::LeaveLFGChannel()
-{
-    for (JoinedChannelsList::iterator i = m_channels.begin(); i != m_channels.end(); ++i)
-    {
-        if ((*i)->IsLFG())
-        {
-            (*i)->Leave(this);
-            break;
-        }
-    }
-}
 
 /**
  * @brief Attempts to improve defense skill and refresh defense-derived bonuses.
